Report model info
Add a script to list the genes with models for a specified model set, along with their batch and number of PCs.

Also updated residual latent pheast to use the newly generated GTEx Pantry phenotypes (using updated annotations and including alt TSS/polyA phenotypes for lncRNA) instead of the converted version of the old phenotypes.